test: add stub test file to ensure all source files get coverage
We're now using `c8`, which doesn't have the `--all` option that `nyc` does... but we can fake it by
including a shim test file whose sole purpose is to import every single source file (that we care
about).  Our apparent coverage has dropped, but that's good; it actually means we're getting a
better picture of the overall coverage.